Fixed _recountLayers to ignore HTML input state
When testing the Control.ActiveLayers class I found that if I added
layers dynamically after creating the control, it would not list them
as active unless I manually toggled the control checkbox.  This returned
unexpected results when I went to retrieve the active overlays later.

It's not clear if this is a race condition due to the HTML element not
being added yet when _recountLayers is being called or if a certain event
is not being handled, but the only way to ensure that active overlays
are properly added seems to be adding all visible layers before creating
the layer control.  Since this isn't likely to happen in many cases,
removing the short-circuit check for the DOM input element seems to
guarantee it's checking the internal map state every time.