Changed _findActiveBaseLayer() to return null instead of error if not found
I changed the behavior of _findActiveBaseLayer() to return null
instead of throwing an error if no baselayer found.  This was
particularly troubling when initializing the Control instance
when no baselayer was initially set.  In some use cases we found
it possible for the map to be initialized with no baselayers and
all layers to be dynamically added later.